+++ ribs/archives/_grid_archive.py
@@ -1,27 +1,44 @@
 """Contains the GridArchive."""
 import numpy as np
 
-from ribs._utils import check_batch_shape, check_finite, check_is_1d, np_scalar
-from ribs.archives._archive_base import ArchiveBase
+from ribs._utils import (check_batch_shape, check_finite, check_is_1d,
+                         check_shape, np_scalar, validate_batch,
+                         validate_single)
+from ribs.archives._archive_base_2 import ArchiveBase
+from ribs.archives._archive_stats import ArchiveStats
+from ribs.archives._array_store import ArrayStore
+from ribs.archives._cqd_score_result import CQDScoreResult
+from ribs.archives._transforms import (batch_entries_with_threshold,
+                                       compute_best_index,
+                                       compute_objective_sum,
+                                       single_entry_with_threshold)
+from ribs.archives._utils import parse_dtype, validate_cma_mae_settings
 
 
 class GridArchive(ArchiveBase):
+    # pylint: disable = too-many-public-methods
     """An archive that divides each dimension into uniformly-sized cells.
 
     This archive is the container described in `Mouret 2015
     <https://arxiv.org/pdf/1504.04909.pdf>`_. It can be visualized as an
     n-dimensional grid in the measure space that is divided into a certain
-    number of cells in each dimension. Each cell contains an elite, i.e. a
-    solution that `maximizes` the objective function for the measures in that
-    cell.
+    number of cells in each dimension. Each cell contains an elite, i.e., a
+    solution that *maximizes* the objective function and has measures that lie
+    within that cell.
 
-    .. note:: The idea of archive thresholds was introduced in `Fontaine 2022
-        <https://arxiv.org/abs/2205.10752>`_. For more info on thresholds,
-        including the ``learning_rate`` and ``threshold_min`` parameters, refer
-        to our tutorial :doc:`/tutorials/cma_mae`.
+    This archive also implements the idea of *soft archives* that have
+    *thresholds*, as introduced in `Fontaine 2023
+    <https://arxiv.org/abs/2205.10752>`_. To learn more about thresholds,
+    including the ``learning_rate`` and ``threshold_min`` parameters, please
+    refer to the tutorial :doc:`/tutorials/cma_mae`.
+
+    By default, this archive stores the following data fields: ``solution``,
+    ``objective``, ``measures``, ``threshold``, and ``index``. The ``threshold``
+    is the value that a solution's objective value must exceed to be inserted
+    into a cell, while the integer ``index`` uniquely identifies each cell.
 
     Args:
-        solution_dim (int): Dimension of the solution space.
+        solution_dim (int): Dimensionality of the solution space.
         dims (array-like of int): Number of cells in each dimension of the
             measure space, e.g. ``[20, 30, 40]`` indicates there should be 3
             dimensions with 20, 30, and 40 cells. (The number of dimensions is
@@ -32,8 +49,8 @@ class GridArchive(ArchiveBase):
             (inclusive), and the second dimension should have bounds
             :math:`[-2,2]` (inclusive). ``ranges`` should be the same length as
             ``dims``.
-        epsilon (float): Due to floating point precision errors, we add a small
-            epsilon when computing the archive indices in the :meth:`index_of`
+        epsilon (float): Due to floating point precision errors, a small epsilon
+            is added when computing the archive indices in the :meth:`index_of`
             method -- refer to the implementation `here
             <../_modules/ribs/archives/_grid_archive.html#GridArchive.index_of>`_.
             Pass this parameter to configure that epsilon.
@@ -52,10 +69,10 @@ class GridArchive(ArchiveBase):
         seed (int): Value to seed the random number generator. Set to None to
             avoid a fixed seed.
         dtype (str or data-type or dict): Data type of the solutions,
-            objectives, and measures. We only support ``"f"`` / ``np.float32``
-            and ``"d"`` / ``np.float64``. Alternatively, this can be a dict
-            specifying separate dtypes, of the form ``{"solution": <dtype>,
-            "objective": <dtype>, "measures": <dtype>}``.
+            objectives, and measures. This can be ``"f"`` / ``np.float32``,
+            ``"d"`` / ``np.float64``, or a dict specifying separate dtypes, of
+            the form ``{"solution": <dtype>, "objective": <dtype>, "measures":
+            <dtype>}``.
         extra_fields (dict): Description of extra fields of data that is stored
             next to elite data like solutions and objectives. The description is
             a dict mapping from a field name (str) to a tuple of ``(shape,
@@ -65,55 +82,145 @@ class GridArchive(ArchiveBase):
             must be valid Python identifiers, and names already used in the
             archive are not allowed.
     Raises:
+        ValueError: Invalid values for learning_rate and threshold_min.
+        ValueError: Invalid names in extra_fields.
         ValueError: ``dims`` and ``ranges`` are not the same length.
     """
 
-    def __init__(self,
-                 *,
-                 solution_dim,
-                 dims,
-                 ranges,
-                 learning_rate=None,
-                 threshold_min=-np.inf,
-                 epsilon=1e-6,
-                 qd_score_offset=0.0,
-                 seed=None,
-                 dtype=np.float64,
-                 extra_fields=None):
+    def __init__(
+        self,
+        *,
+        solution_dim,
+        dims,
+        ranges,
+        learning_rate=None,
+        threshold_min=-np.inf,
+        epsilon=1e-6,
+        qd_score_offset=0.0,
+        seed=None,
+        dtype=np.float64,
+        extra_fields=None,
+    ):
+        self._rng = np.random.default_rng(seed)
         self._dims = np.array(dims, dtype=np.int32)
-        if len(self._dims) != len(ranges):
-            raise ValueError(f"dims (length {len(self._dims)}) and ranges "
-                             f"(length {len(ranges)}) must be the same length")
 
         ArchiveBase.__init__(
             self,
             solution_dim=solution_dim,
-            cells=np.prod(self._dims),
+            objective_dim=(),
             measure_dim=len(self._dims),
-            learning_rate=learning_rate,
-            threshold_min=threshold_min,
-            qd_score_offset=qd_score_offset,
-            seed=seed,
-            dtype=dtype,
-            extra_fields=extra_fields,
         )
 
-        ranges = list(zip(*ranges))
+        # Set up the ArrayStore, which is a data structure that stores all the
+        # elites' data in arrays sharing a common index.
+        extra_fields = extra_fields or {}
+        reserved_fields = {
+            "solution", "objective", "measures", "threshold", "index"
+        }
+        if reserved_fields & extra_fields.keys():
+            raise ValueError("The following names are not allowed in "
+                             f"extra_fields: {reserved_fields}")
+        dtype = parse_dtype(dtype)
+        self._store = ArrayStore(
+            field_desc={
+                "solution": ((self.solution_dim,), dtype["solution"]),
+                "objective": ((), dtype["objective"]),
+                "measures": ((self.measure_dim,), dtype["measures"]),
+                # Must be same dtype as the objective since they share
+                # calculations.
+                "threshold": ((), dtype["objective"]),
+                **extra_fields,
+            },
+            capacity=np.prod(self._dims),
+        )
+
+        # Set up constant properties.
+        if len(self._dims) != len(ranges):
+            raise ValueError(f"dims (length {len(self._dims)}) and ranges "
+                             f"(length {len(ranges)}) must be the same length")
+        ranges = list(zip(*ranges))  # Rearrange into lower and upper bounds.
         self._lower_bounds = np.array(ranges[0], dtype=self.dtypes["measures"])
         self._upper_bounds = np.array(ranges[1], dtype=self.dtypes["measures"])
         self._interval_size = self._upper_bounds - self._lower_bounds
+        self._boundaries = self._compute_boundaries(self._dims,
+                                                    self._lower_bounds,
+                                                    self._upper_bounds)
         self._epsilon = np_scalar(epsilon, dtype=self.dtypes["measures"])
+        self._learning_rate, self._threshold_min = validate_cma_mae_settings(
+            learning_rate, threshold_min, self.dtypes["threshold"])
+        self._qd_score_offset = np_scalar(qd_score_offset,
+                                          self.dtypes["objective"])
 
-        self._boundaries = []
-        for dim, lower_bound, upper_bound in zip(self._dims, self._lower_bounds,
-                                                 self._upper_bounds):
-            self._boundaries.append(
-                np.linspace(lower_bound, upper_bound, dim + 1))
+        # Set up statistics.
+        self._stats = None
+        self._best_elite = None
+        # Sum of all objective values in the archive; useful for computing
+        # qd_score and obj_mean.
+        self._objective_sum = None
+        self._stats_reset()
+
+    @staticmethod
+    def _compute_boundaries(dims, lower_bounds, upper_bounds):
+        """Computes grid cell boundaries of the archive."""
+        boundaries = []
+        for dim, lower_bound, upper_bound in zip(dims, lower_bounds,
+                                                 upper_bounds):
+            boundaries.append(np.linspace(lower_bound, upper_bound, dim + 1))
+        return boundaries
+
+    ## Properties inherited from ArchiveBase ##
+
+    @property
+    def field_list(self):
+        return self._store.field_list
+
+    @property
+    def stats(self):
+        return self._stats
+
+    @property
+    def empty(self):
+        return len(self._store) == 0
+
+    @property
+    def dtypes(self):
+        return self._store.dtypes
+
+    ## Properties that are not in ArchiveBase ##
+    ## Roughly ordered by the parameter list in the constructor. ##
+
+    @property
+    def best_elite(self):
+        """dict: The elite with the highest objective in the archive.
+
+        None if there are no elites in the archive.
+
+        .. note::
+            If the archive is non-elitist (this occurs when using the archive
+            with a learning rate which is not 1.0, as in CMA-MAE), then this
+            best elite may no longer exist in the archive because it was
+            replaced with an elite with a lower objective value. This can happen
+            because in non-elitist archives, new solutions only need to exceed
+            the *threshold* of the cell they are being inserted into, not the
+            *objective* of the elite currently in the cell. See :pr:`314` for
+            more info.
+
+        .. note::
+            The best elite will contain a "threshold" key. This threshold is the
+            threshold of the best elite's cell after the best elite was inserted
+            into the archive.
+        """
+        return self._best_elite
 
     @property
     def dims(self):
         """(measure_dim,) numpy.ndarray: Number of cells in each dimension."""
         return self._dims
+
+    @property
+    def cells(self):
+        """int: Total number of cells in the archive."""
+        return self._store.capacity
 
     @property
     def lower_bounds(self):
@@ -130,12 +237,6 @@ class GridArchive(ArchiveBase):
         """(measure_dim,) numpy.ndarray: The size of each dim (upper_bounds -
         lower_bounds)."""
         return self._interval_size
-
-    @property
-    def epsilon(self):
-        """dtypes["measures"]: Epsilon for computing archive indices. Refer to
-        the documentation for this class."""
-        return self._epsilon
 
     @property
     def boundaries(self):
@@ -155,25 +256,99 @@ class GridArchive(ArchiveBase):
         """
         return self._boundaries
 
+    @property
+    def epsilon(self):
+        """dtypes["measures"]: Epsilon for computing archive indices. Refer to
+        the documentation for this class."""
+        return self._epsilon
+
+    @property
+    def learning_rate(self):
+        """float: The learning rate for threshold updates."""
+        return self._learning_rate
+
+    @property
+    def threshold_min(self):
+        """float: The initial threshold value for all the cells."""
+        return self._threshold_min
+
+    @property
+    def qd_score_offset(self):
+        """float: The offset which is subtracted from objective values when
+        computing the QD score."""
+        return self._qd_score_offset
+
+    ## dunder methods ##
+
+    def __len__(self):
+        return len(self._store)
+
+    def __iter__(self):
+        return iter(self._store)
+
+    ## Utilities ##
+
+    def _stats_reset(self):
+        """Resets the archive stats."""
+        zero = np_scalar(0.0, dtype=self.dtypes["objective"])
+
+        self._stats = ArchiveStats(
+            num_elites=0,
+            coverage=zero,
+            qd_score=zero,
+            norm_qd_score=zero,
+            obj_max=None,
+            obj_mean=None,
+        )
+        self._best_elite = None
+        self._objective_sum = zero
+
+    def _stats_update(self, new_objective_sum, new_best_index):
+        """Updates statistics based on a new sum of objective values
+        (new_objective_sum) and the index of a potential new best elite
+        (new_best_index)."""
+        self._objective_sum = new_objective_sum
+        new_qd_score = (self._objective_sum -
+                        np_scalar(len(self), dtype=self.dtypes["objective"]) *
+                        self._qd_score_offset)
+
+        _, new_best_elite = self._store.retrieve([new_best_index])
+
+        if (self._stats.obj_max is None or
+                new_best_elite["objective"] > self._stats.obj_max):
+            # Convert batched values to single values.
+            new_best_elite = {k: v[0] for k, v in new_best_elite.items()}
+
+            new_obj_max = new_best_elite["objective"]
+            self._best_elite = new_best_elite
+        else:
+            new_obj_max = self._stats.obj_max
+
+        self._stats = ArchiveStats(
+            num_elites=len(self),
+            coverage=np_scalar(len(self) / self.cells,
+                               dtype=self.dtypes["objective"]),
+            qd_score=new_qd_score,
+            norm_qd_score=np_scalar(new_qd_score / self.cells,
+                                    dtype=self.dtypes["objective"]),
+            obj_max=new_obj_max,
+            obj_mean=np_scalar(self._objective_sum / len(self),
+                               dtype=self.dtypes["objective"]),
+        )
+
     def index_of(self, measures):
         """Returns archive indices for the given batch of measures.
 
         First, values are clipped to the bounds of the measure space. Then, the
-        values are mapped to cells; e.g. cell 5 along dimension 0 and cell 3
-        along dimension 1.
+        values are mapped to indices in the grid, e.g., cell 5 along dimension 0
+        and cell 3 along dimension 1. We convert these grid indices to integer
+        indices with :func:`numpy.ravel_multi_index` and return the result.
 
-        At this point, we have "grid indices" -- indices of each measure in each
-        dimension. Since indices returned by this method must be single integers
-        (as opposed to a tuple of grid indices), we convert these grid indices
-        into integer indices with :func:`numpy.ravel_multi_index` and return the
-        result.
+        To convert between integer and grid indices, we also provide utility
+        methods :meth:`grid_to_int_index` and :meth:`int_to_grid_index`.
 
-        It may be useful to have the original grid indices. Thus, we provide the
-        :meth:`grid_to_int_index` and :meth:`int_to_grid_index` methods for
-        converting between grid and integer indices.
-
-        As an example, the grid indices can be used to access boundaries of a
-        measure value's cell. For example, the following retrieves the lower
+        As an example, grid indices can be used to access :attr:`boundaries` of
+        a measure value's cell. For example, the following retrieves the lower
         and upper bounds of the cell along dimension 0::
 
             # Access only element 0 since this method operates in batch.
@@ -181,8 +356,6 @@ class GridArchive(ArchiveBase):
 
             lower = archive.boundaries[0][idx[0]]
             upper = archive.boundaries[0][idx[0] + 1]
-
-        See :attr:`boundaries` for more info.
 
         Args:
             measures (array-like): (batch_size, :attr:`measure_dim`) array of
@@ -211,6 +384,26 @@ class GridArchive(ArchiveBase):
         grid_indices = np.clip(grid_indices, 0, self._dims - 1)
 
         return self.grid_to_int_index(grid_indices)
+
+    def index_of_single(self, measures):
+        """Returns the index of the measures for one solution.
+
+        See :meth:`index_of`.
+
+        Args:
+            measures (array-like): (:attr:`measure_dim`,) array of measures for
+                a single solution.
+        Returns:
+            int or numpy.integer: Integer index of the measures in the archive's
+            storage arrays.
+        Raises:
+            ValueError: ``measures`` is not of shape (:attr:`measure_dim`,).
+            ValueError: ``measures`` has non-finite values (inf or NaN).
+        """
+        measures = np.asarray(measures)
+        check_shape(measures, "measures", self.measure_dim, "measure_dim")
+        check_finite(measures, "measures")
+        return self.index_of(measures[None])[0]
 
     def grid_to_int_index(self, grid_indices):
         """Converts a batch of grid indices into a batch of integer indices.
@@ -253,3 +446,389 @@ class GridArchive(ArchiveBase):
             int_indices,
             self._dims,
         )).T.astype(np.int32)
+
+    ## Methods for writing to the archive ##
+
+    def add(self, solution, objective, measures, **fields):
+        """Inserts a batch of solutions into the archive.
+
+        Each solution is only inserted if it has a higher ``objective`` than the
+        threshold of the corresponding cell. For the default values of
+        ``learning_rate`` and ``threshold_min``, this threshold is simply the
+        objective value of the elite previously in the cell.  If multiple
+        solutions in the batch end up in the same cell, we only insert the
+        solution with the highest objective. If multiple solutions end up in the
+        same cell and tie for the highest objective, we insert the solution that
+        appears first in the batch.
+
+        For the default values of ``learning_rate`` and ``threshold_min``, the
+        threshold for each cell is updated by taking the maximum objective value
+        among all the solutions that landed in the cell, resulting in the same
+        behavior as in the vanilla MAP-Elites archive. However, for other
+        settings, the threshold is updated with the batch update rule described
+        in the appendix of `Fontaine 2023 <https://arxiv.org/abs/2205.10752>`_.
+
+        .. note:: The indices of all arguments should "correspond" to each
+            other, i.e., ``solution[i]``, ``objective[i]``,
+            ``measures[i]``, and should be the solution parameters,
+            objective, and measures for solution ``i``.
+
+        Args:
+            solution (array-like): (batch_size, :attr:`solution_dim`) array of
+                solution parameters.
+            objective (array-like): (batch_size,) array with objective function
+                evaluations of the solutions.
+            measures (array-like): (batch_size, :attr:`measure_dim`) array with
+                measure space coordinates of all the solutions.
+            fields (keyword arguments): Additional data for each solution. Each
+                argument should be an array with batch_size as the first
+                dimension.
+
+        Returns:
+            dict: Information describing the result of the add operation. The
+            dict contains the following keys:
+
+            - ``"status"`` (:class:`numpy.ndarray` of :class:`int`): An array of
+              integers that represent the "status" obtained when attempting to
+              insert each solution in the batch. Each item has the following
+              possible values:
+
+              - ``0``: The solution was not added to the archive.
+              - ``1``: The solution improved the objective value of a cell
+                which was already in the archive.
+              - ``2``: The solution discovered a new cell in the archive.
+
+              All statuses (and values, below) are computed with respect to the
+              *current* archive. For example, if two solutions both introduce
+              the same new archive cell, then both will be marked with ``2``.
+
+              The alternative is to depend on the order of the solutions in the
+              batch -- for example, if we have two solutions ``a`` and ``b``
+              that introduce the same new cell in the archive, ``a`` could be
+              inserted first with status ``2``, and ``b`` could be inserted
+              second with status ``1`` because it improves upon ``a``. However,
+              our implementation does **not** do this.
+
+              To convert statuses to a more semantic format, cast all statuses
+              to :class:`AddStatus`, e.g., with ``[AddStatus(s) for s in
+              add_info["status"]]``.
+
+            - ``"value"`` (:class:`numpy.ndarray` of
+              :attr:`dtypes` ["objective"]): An array with values for each
+              solution in the batch. With the default values of ``learning_rate
+              = 1.0`` and ``threshold_min = -np.inf``, the meaning of each value
+              depends on the corresponding ``status`` and is identical to that
+              in CMA-ME (`Fontaine 2020 <https://arxiv.org/abs/1912.02400>`_):
+
+              - ``0`` (not added): The value is the "negative improvement,"
+                i.e., the objective of the solution passed in minus the
+                objective of the elite still in the archive (this value is
+                negative because the solution did not have a high enough
+                objective to be added to the archive).
+              - ``1`` (improve existing cell): The value is the "improvement,"
+                i.e., the objective of the solution passed in minus the
+                objective of the elite previously in the archive.
+              - ``2`` (new cell): The value is just the objective of the
+                solution.
+
+              In contrast, for other values of ``learning_rate`` and
+              ``threshold_min``, each value is equivalent to the objective value
+              of the solution minus the threshold of its corresponding cell in
+              the archive.
+
+        Raises:
+            ValueError: The array arguments do not match their specified shapes.
+            ValueError: ``objective`` or ``measures`` has non-finite values (inf
+                or NaN).
+        """
+        data = validate_batch(
+            self,
+            {
+                "solution": solution,
+                "objective": objective,
+                "measures": measures,
+                **fields,
+            },
+        )
+
+        add_info = self._store.add(
+            self.index_of(data["measures"]),
+            data,
+            {
+                "dtype": self.dtypes["threshold"],
+                "learning_rate": self._learning_rate,
+                "threshold_min": self._threshold_min,
+                "objective_sum": self._objective_sum,
+            },
+            [
+                batch_entries_with_threshold,
+                compute_objective_sum,
+                compute_best_index,
+            ],
+        )
+
+        objective_sum = add_info.pop("objective_sum")
+        best_index = add_info.pop("best_index")
+        if not np.all(add_info["status"] == 0):
+            self._stats_update(objective_sum, best_index)
+
+        return add_info
+
+    def add_single(self, solution, objective, measures, **fields):
+        """Inserts a single solution into the archive.
+
+        The solution is only inserted if it has a higher ``objective`` than the
+        threshold of the corresponding cell. For the default values of
+        ``learning_rate`` and ``threshold_min``, this threshold is simply the
+        objective value of the elite previously in the cell.  The threshold is
+        also updated if the solution was inserted.
+
+        .. note::
+            To make it more amenable to modifications, this method's
+            implementation is designed to be readable at the cost of
+            performance, e.g., none of its operations are modified. :meth:`add`
+            should be used if performance is required.
+
+        Args:
+            solution (array-like): Parameters of the solution.
+            objective (float): Objective function evaluation of the solution.
+            measures (array-like): Coordinates in measure space of the solution.
+            fields (keyword arguments): Additional data for the solution.
+
+        Returns:
+            dict: Information describing the result of the add operation. The
+            dict contains ``status`` and ``value`` keys; refer to :meth:`add`
+            for the meaning of status and value.
+
+        Raises:
+            ValueError: The array arguments do not match their specified shapes.
+            ValueError: ``objective`` is non-finite (inf or NaN) or ``measures``
+                has non-finite values.
+        """
+        data = validate_single(
+            self,
+            {
+                "solution": solution,
+                "objective": objective,
+                "measures": measures,
+                **fields,
+            },
+        )
+
+        for name, arr in data.items():
+            data[name] = np.expand_dims(arr, axis=0)
+
+        add_info = self._store.add(
+            np.expand_dims(self.index_of_single(measures), axis=0),
+            data,
+            {
+                "dtype": self.dtypes["threshold"],
+                "learning_rate": self._learning_rate,
+                "threshold_min": self._threshold_min,
+                "objective_sum": self._objective_sum,
+            },
+            [
+                single_entry_with_threshold,
+                compute_objective_sum,
+                compute_best_index,
+            ],
+        )
+
+        objective_sum = add_info.pop("objective_sum")
+        best_index = add_info.pop("best_index")
+
+        for name, arr in add_info.items():
+            add_info[name] = arr[0]
+
+        if add_info["status"]:
+            self._stats_update(objective_sum, best_index)
+
+        return add_info
+
+    def clear(self):
+        """Removes all elites in the archive."""
+        self._store.clear()
+        self._stats_reset()
+
+    ## Methods for reading from the archive ##
+    ## Refer to ArchiveBase for documentation of these methods. ##
+
+    def retrieve(self, measures):
+        measures = np.asarray(measures)
+        check_batch_shape(measures, "measures", self.measure_dim, "measure_dim")
+        check_finite(measures, "measures")
+
+        occupied, data = self._store.retrieve(self.index_of(measures))
+        unoccupied = ~occupied
+
+        for name, arr in data.items():
+            if arr.dtype == object:
+                fill_val = None
+            elif name == "index":
+                fill_val = -1
+            elif np.issubdtype(arr.dtype, np.integer):
+                fill_val = 0
+            else:  # Floating-point and other fields.
+                fill_val = np.nan
+
+            arr[unoccupied] = fill_val
+
+        return occupied, data
+
+    def retrieve_single(self, measures):
+        measures = np.asarray(measures)
+        check_shape(measures, "measures", self.measure_dim, "measure_dim")
+        check_finite(measures, "measures")
+
+        occupied, data = self.retrieve(measures[None])
+
+        return occupied[0], {field: arr[0] for field, arr in data.items()}
+
+    def data(self, fields=None, return_type="dict"):
+        return self._store.data(fields, return_type)
+
+    def sample_elites(self, n):
+        if self.empty:
+            raise IndexError("No elements in archive.")
+
+        random_indices = self._rng.integers(len(self._store), size=n)
+        selected_indices = self._store.occupied_list[random_indices]
+        _, elites = self._store.retrieve(selected_indices)
+        return elites
+
+    ## CQD Score ##
+
+    def cqd_score(self,
+                  iterations,
+                  target_points,
+                  penalties,
+                  obj_min,
+                  obj_max,
+                  dist_max=None,
+                  dist_ord=None):
+        """Computes the CQD score of the archive.
+
+        The Continuous Quality Diversity (CQD) score was introduced in
+        `Kent 2022 <https://dl.acm.org/doi/10.1145/3520304.3534018>`_.
+
+        .. note:: This method by default assumes that the archive has an
+            ``upper_bounds`` and ``lower_bounds`` property which delineate the
+            bounds of the measure space, as is the case in
+            :class:`~ribs.archives.GridArchive`,
+            :class:`~ribs.archives.CVTArchive`, and
+            :class:`~ribs.archives.SlidingBoundariesArchive`.  If this is not
+            the case, ``dist_max`` must be passed in, and ``target_points`` must
+            be an array of custom points.
+
+        Args:
+            iterations (int): Number of times to compute the CQD score. The mean
+                CQD score across these iterations is returned.
+            target_points (int or array-like): Number of target points to
+                generate, or an (iterations, n, measure_dim) array which
+                lists n target points to list on each iteration. When an int is
+                passed, the points are sampled uniformly within the bounds of
+                the measure space.
+            penalties (int or array-like): Number of penalty values over which
+                to compute the score (the values are distributed evenly over the
+                range [0,1]). Alternatively, this may be a 1D array which
+                explicitly lists the penalty values. Known as :math:`\\theta` in
+                Kent 2022.
+            obj_min (float): Minimum objective value, used when normalizing the
+                objectives.
+            obj_max (float): Maximum objective value, used when normalizing the
+                objectives.
+            dist_max (float): Maximum distance between points in measure space.
+                Defaults to the distance between the extremes of the measure
+                space bounds (the type of distance is computed with the order
+                specified by ``dist_ord``). Known as :math:`\\delta_{max}` in
+                Kent 2022.
+            dist_ord: Order of the norm to use for calculating measure space
+                distance; this is passed to :func:`numpy.linalg.norm` as the
+                ``ord`` argument. See :func:`numpy.linalg.norm` for possible
+                values. The default is to use Euclidean distance (L2 norm).
+        Returns:
+            The mean CQD score obtained with ``iterations`` rounds of
+            calculations.
+        Raises:
+            RuntimeError: The archive does not have the bounds properties
+                mentioned above, and dist_max is not specified or the target
+                points are not provided.
+            ValueError: target_points or penalties is an array with the wrong
+                shape.
+        """
+        if (not (hasattr(self, "upper_bounds") and
+                 hasattr(self, "lower_bounds")) and
+            (dist_max is None or np.isscalar(target_points))):
+            raise RuntimeError(
+                "When the archive does not have lower_bounds and "
+                "upper_bounds properties, dist_max must be specified, "
+                "and target_points must be an array")
+
+        if np.isscalar(target_points):
+            # pylint: disable = no-member
+            target_points = self._rng.uniform(
+                low=self.lower_bounds,
+                high=self.upper_bounds,
+                size=(iterations, target_points, self.measure_dim),
+            )
+        else:
+            # Copy since this is returned.
+            target_points = np.copy(target_points)
+            if (target_points.ndim != 3 or
+                    target_points.shape[0] != iterations or
+                    target_points.shape[2] != self.measure_dim):
+                raise ValueError(
+                    "Expected target_points to be a 3D array with "
+                    f"shape ({iterations}, n, {self.measure_dim}) "
+                    "(i.e. shape (iterations, n, measure_dim)) but it had "
+                    f"shape {target_points.shape}")
+
+        if dist_max is None:
+            # pylint: disable = no-member
+            dist_max = np.linalg.norm(self.upper_bounds - self.lower_bounds,
+                                      ord=dist_ord)
+
+        if np.isscalar(penalties):
+            penalties = np.linspace(0, 1, penalties)
+        else:
+            penalties = np.copy(penalties)  # Copy since this is returned.
+            check_is_1d(penalties, "penalties")
+
+        objective_batch = self._store.data("objective")
+        measures_batch = self._store.data("measures")
+
+        norm_objectives = objective_batch / (obj_max - obj_min)
+
+        scores = np.zeros(iterations)
+
+        for itr in range(iterations):
+            # Distance calculation -- start by taking the difference between
+            # each measure i and all the target points.
+            distances = measures_batch[:, None] - target_points[itr]
+
+            # (len(archive), n_target_points) array of distances.
+            distances = np.linalg.norm(distances, ord=dist_ord, axis=2)
+
+            norm_distances = distances / dist_max
+
+            for penalty in penalties:
+                # Known as omega in Kent 2022 -- a (len(archive),
+                # n_target_points) array.
+                values = norm_objectives[:, None] - penalty * norm_distances
+
+                # (n_target_points,) array.
+                max_values_per_target = np.max(values, axis=0)
+
+                scores[itr] += np.sum(max_values_per_target)
+
+        return CQDScoreResult(
+            iterations=iterations,
+            mean=np.mean(scores),
+            scores=scores,
+            target_points=target_points,
+            penalties=penalties,
+            obj_min=obj_min,
+            obj_max=obj_max,
+            dist_max=dist_max,
+            dist_ord=dist_ord,
+        )