+++ HISTORY.md
@@ -14,9 +14,10 @@
 - Add DensityRanker for density descent search ({pr}`483`, {pr}`504`)
 - Add NoveltyRanker for novelty search ({pr}`477`)
 - Add proximity_archive_plot for visualizing ProximityArchive ({pr}`476`,
-  {pr}`480`)
+  {pr}`480`, {pr}`523`)
 - Support novelty search with local competition in ProximityArchive ({pr}`481`)
-- Add ProximityArchive for novelty search ({pr}`472`, {pr}`479`, {pr}`484`)
+- Add ProximityArchive for novelty search ({pr}`472`, {pr}`479`, {pr}`484`,
+  {pr}`521`)
 - Support diversity optimization in Scheduler.tell ({pr}`473`)
 - Allow specifying separate dtypes for solution, objective, and measures
   ({pr}`471`)
@@ -25,6 +26,7 @@
 
 #### Bugs
 
+- Make emitter bounds dtype match solution dtype ({pr}`519`)
 - Fix `BanditScheduler` behaviour: the number of active emitters remains stable
   ({pr}`489`)
 
@@ -42,8 +44,11 @@
 
 #### Improvements
 
+- Refactor archives into single-file implementations ({pr}`518`, {pr}`521`)
+- Make ArrayStore.data return ArchiveDataFrame instead of DataFrame ({pr}`522`)
 - Migrate to pyproject.toml ({pr}`514`)
-- Set vmin and vmax to None if archive is empty in ribs.visualize ({pr}`513`)
+- Set vmin and vmax to None if archive is empty in ribs.visualize ({pr}`513`,
+  {pr}`523`)
 - Remove operators from GaussianEmitter and IsoLineEmitter ({pr}`508`)
 - Update QDax visualizations to match QDax 0.5.0 ({pr}`502`)
 - Skip qdax tests if qdax not installed ({pr}`491`)